print-monitor/requirements.txt
argparse==1.2.1
lxml==3.3.0
python-rrdtool==1.4.7
requests==2.2.1
wsgiref==0.1.2
